+++ Simple-Configuration/src/main/java/org/simpleyaml/configuration/comments/format/CommentFormatterConfiguration.java
@@ -1,4 +1,4 @@
-package org.simpleyaml.configuration.comments;
+package org.simpleyaml.configuration.comments.format;
 
 import org.simpleyaml.utils.StringUtils;
