+++ Simple-Yaml/src/main/java/org/simpleyaml/configuration/comments/format/YamlCommentFormat.java
@@ -1,4 +1,4 @@
-package org.simpleyaml.configuration.comments;
+package org.simpleyaml.configuration.comments.format;
 
 public enum YamlCommentFormat {
