+++ Simple-Yaml/src/main/java/org/simpleyaml/configuration/comments/format/YamlCommentFormatterConfiguration.java
@@ -1,4 +1,4 @@
-package org.simpleyaml.configuration.comments;
+package org.simpleyaml.configuration.comments.format;
 
 import org.simpleyaml.utils.StringUtils;
 import org.simpleyaml.utils.Validate;
