+++ Simple-Yaml/src/main/java/org/simpleyaml/configuration/implementation/snakeyaml/SnakeYamlQuoteValue.java
@@ -1,33 +1,19 @@
-package org.simpleyaml.configuration.implementation;
+package org.simpleyaml.configuration.implementation.snakeyaml;
 
 import org.simpleyaml.configuration.implementation.api.QuoteStyle;
 import org.simpleyaml.configuration.implementation.api.QuoteValue;
-import org.simpleyaml.utils.StringUtils;
 import org.yaml.snakeyaml.DumperOptions;
 
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Objects;
 
-public class SnakeYamlQuoteValue<T> implements QuoteValue<T> {
+public class SnakeYamlQuoteValue<T> extends QuoteValue<T> {
 
     private static final Map<QuoteStyle, DumperOptions.ScalarStyle> QUOTE_SCALAR_STYLES = mapQuoteScalarStyles();
 
-    private final T value;
-    private final QuoteStyle quoteStyle;
-
     public SnakeYamlQuoteValue(T value, QuoteStyle quoteStyle) {
-        this.value = value;
-        this.quoteStyle = quoteStyle;
-    }
-
-    public T getValue() {
-        return this.value;
-    }
-
-    public QuoteStyle getQuoteStyle() {
-        return this.quoteStyle;
+        super(value, quoteStyle);
     }
 
     public static DumperOptions.ScalarStyle getQuoteScalarStyle(final QuoteStyle quoteStyle) {
@@ -41,23 +27,5 @@ public class SnakeYamlQuoteValue<T> implements QuoteValue<T> {
         map.put(QuoteStyle.PLAIN, DumperOptions.ScalarStyle.PLAIN);
         map.put(null, DumperOptions.ScalarStyle.PLAIN);
         return Collections.unmodifiableMap(map);
-    }
-
-    @Override
-    public String toString() {
-        return this.quoteStyle.toString() + "=" + (this.value == null ? "!!null" : StringUtils.quoteNewLines(this.value.toString()));
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        SnakeYamlQuoteValue<?> that = (SnakeYamlQuoteValue<?>) o;
-        return Objects.equals(value, that.value) && quoteStyle == that.quoteStyle;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(value, quoteStyle);
     }
 }
