+++ Simple-Yaml/src/test/java/org/simpleyaml/configuration/file/YamlConfigurationOptionsTest.java
@@ -7,8 +7,8 @@ import org.junit.jupiter.api.Test;
 import org.llorllale.cactoos.matchers.IsTrue;
 import org.llorllale.cactoos.matchers.Throws;
 import org.simpleyaml.configuration.comments.CommentType;
-import org.simpleyaml.configuration.comments.YamlCommentFormat;
-import org.simpleyaml.configuration.comments.YamlCommentFormatterConfiguration;
+import org.simpleyaml.configuration.comments.format.YamlCommentFormat;
+import org.simpleyaml.configuration.comments.format.YamlCommentFormatterConfiguration;
 import org.simpleyaml.configuration.implementation.api.QuoteStyle;
 
 import java.io.IOException;
