+++ Simple-Yaml/src/test/java/org/simpleyaml/configuration/file/YamlFileTest.java
@@ -8,7 +8,11 @@ import org.hamcrest.core.IsNull;
 import org.hamcrest.core.IsSame;
 import org.junit.jupiter.api.Test;
 import org.llorllale.cactoos.matchers.IsTrue;
-import org.simpleyaml.configuration.comments.*;
+import org.simpleyaml.configuration.comments.CommentType;
+import org.simpleyaml.configuration.comments.format.YamlCommentFormat;
+import org.simpleyaml.configuration.comments.format.YamlCommentFormatter;
+import org.simpleyaml.configuration.comments.format.YamlCommentFormatterConfiguration;
+import org.simpleyaml.configuration.comments.format.YamlHeaderFormatter;
 import org.simpleyaml.configuration.implementation.api.QuoteStyle;
 import org.simpleyaml.examples.Person;
 import org.simpleyaml.obj.TestResources;
